Create readme.md
For the IBM Data Science Capstone Project we used rocket launch data from SpaceX to analyze and determine the factors that best lead to a successful rocket launch. 

All Jupyter notebooks that were used to perform the analysis are included in this folder.

The final deliverable is a report summarizing the findings.